+++ include/fractals/common/utils.h
@@ -82,9 +82,9 @@ namespace fractals::common {
     std::vector<bool> bytes_to_bitfield(int len,std::deque<char> &bitfields);
     std::vector<bool> bytes_to_bitfield(int len,std::vector<char> &bitfields);
 
-    std::wstring pp_bytes(long long bytes);
-    std::wstring pp_bytes_per_second(long long bytes);
-    std::wstring pp_time(long long seconds);
+    std::wstring pp_bytes(int64_t bytes);
+    std::wstring pp_bytes_per_second(int64_t bytes);
+    std::wstring pp_time(int64_t seconds);
     std::wstring make_wide(const std::string &s);
     std::string unwide(const std::wstring &ws);
